extract receipt service to own file
