+++ Yosemite/Yosemite/Tools/POS/POSOrderService.swift
@@ -88,39 +88,6 @@ public protocol POSOrderServiceProtocol {
     func updatePOSOrder(order: Order, recipientEmail: String) async throws
 }
 
-public protocol POSReceiptServiceProtocol {
-    func sendReceipt(order: Order, recipientEmail: String) async throws
-}
-
-public final class POSReceiptService: POSReceiptServiceProtocol {
-    private let siteID: Int64
-    private let receiptsRemote: POSReceiptsRemoteProtocol
-
-    public convenience init?(siteID: Int64, credentials: Credentials?) {
-        guard let credentials else {
-            DDLogError("⛔️ Could not create POSReceiptService due to not finding credentials")
-            return nil
-        }
-        let network = AlamofireNetwork(credentials: credentials)
-        self.init(siteID: siteID,
-                  receiptsRemote: ReceiptRemote(network: network))
-    }
-
-    public init(siteID: Int64,
-                receiptsRemote: POSReceiptsRemoteProtocol) {
-        self.siteID = siteID
-        self.receiptsRemote = receiptsRemote
-    }
-
-    public func sendReceipt(order: Yosemite.Order, recipientEmail: String) async throws {
-        do {
-            try await receiptsRemote.sendReceipt(siteID: siteID, orderID: order.orderID)
-        } catch {
-            throw POSReceiptServiceError.sendReceiptFailed
-        }
-    }
-}
-
 public final class POSOrderService: POSOrderServiceProtocol {
     private let siteID: Int64
     private let ordersRemote: POSOrdersRemoteProtocol
@@ -224,11 +191,5 @@ private extension POSOrderService {
     enum POSOrderServiceError: Error {
         case emailAlreadySet
         case updateOrderFailed
-    }
-}
-
-private extension POSReceiptService {
-    enum POSReceiptServiceError: Error {
-        case sendReceiptFailed
     }
 }
